test: organize retained model tests by family
