+++ tests/unit/train/models/qwen3/test_model.py
@@ -122,33 +122,3 @@ def test_model_with_sequence_packing(model, correct_position_ids):
         torch.testing.assert_close(output_packed_left, output_base)
         with pytest.raises(AssertionError):
             torch.testing.assert_close(output_packed_right, output_base)
-
-
-def test_moe_custom_impl():
-    config = ModelConfig(name="PrimeIntellect/GLM-0.5B", attn="flash_attention_2", impl="custom")
-    model = get_model(config)
-    model = model.to("cuda")
-    # we need to wrap the lm head as custom forward only works with it, this is done in setup_model
-    inject_prime_lm_head(model, chunk_size=None)
-    with torch.autocast("cuda", dtype=torch.bfloat16):
-        inputs_ids = torch.randint(0, 100, (BS, SEQ_LEN)).to("cuda")
-        outputs = model(input_ids=inputs_ids, seq_lens=torch.tensor([SEQ_LEN], device="cuda"))
-        logits = outputs["logits"]
-
-        assert logits.shape == (BS, SEQ_LEN, model.config.vocab_size)
-
-
-@pytest.mark.skip(reason="need special token for meta stuff in ci")
-@pytest.mark.parametrize("model_name", ["meta-llama/Llama-3.2-1B-Instruct"])
-def test_model_forward_custom_impl(model_name):
-    config = ModelConfig(name=model_name, impl="custom", attn="flash_attention_2")
-    model = get_model(config)
-    # we need to wrap the lm head as custom forward only works with it, this is done in setup_model
-    inject_prime_lm_head(model, chunk_size=None)
-    model = model.to("cuda")
-    with torch.autocast("cuda", dtype=torch.bfloat16):
-        inputs_ids = torch.randint(0, 100, (BS, SEQ_LEN)).to("cuda")
-        outputs = model(input_ids=inputs_ids, seq_lens=torch.tensor([SEQ_LEN], device="cuda"))
-        logits = outputs["logits"]
-
-        assert logits.shape == (BS, SEQ_LEN, model.config.vocab_size)
