+++ tests/unit/train/models/qwen3_moe/test_model.py
